+++ app/src/main/java/io/github/gsantner/memetastic/ui/GridRecycleAdapter.java
@@ -115,7 +115,7 @@ public class GridRecycleAdapter extends RecyclerView.Adapter<GridRecycleAdapter.
 
     private void tintFavourite(ImageView iv, boolean isFav) {
         Helpers.get().setDrawableWithColorToImageView(iv,
-                isFav ? R.drawable.ic_star_black_48px : R.drawable.ic_star_border_black_48px,
+                isFav ? R.drawable.ic_star_black_32dp : R.drawable.ic_star_border_black_32dp,
                 isFav ? R.color.comic_yellow : R.color.comic_blue);
     }