+++ app/src/main/java/io/github/gsantner/memetastic/util/AppSettings.java
@@ -71,7 +71,7 @@ public class AppSettings extends AppSettingsBase {
     }
 
     public int getRenderQualityReal() {
-        int val = getInt(prefApp, R.string.pref_key__render_quality_editor_percent, 24);
+        int val = getInt(R.string.pref_key__render_quality_editor_percent, 24);
         return (int) (400 + (2100.0 * (val / 100.0)));
     }
 
@@ -80,24 +80,24 @@ public class AppSettings extends AppSettingsBase {
         // Additional reduction of quality to ~2/3 is roughly 225
         // 150 is very fast loaded, but blurry, 200 is still a little blurry, 225 seems to be
         // a good tradeoff between quality (400-600) and speed (-125)
-        int val = getInt(prefApp, R.string.pref_key__thumbnail_quality__percent, 24);
+        int val = getInt(R.string.pref_key__thumbnail_quality__percent, 24);
         return (int) (100 + (939 * (val / 100.0)));
     }
 
     public void setLastSelectedFont(int value) {
-        setInt(prefApp, R.string.pref_key__last_selected_font, value);
+        setInt(R.string.pref_key__last_selected_font, value);
     }
 
     public int getLastSelectedFont() {
-        return getInt(prefApp, R.string.pref_key__last_selected_font, 0);
+        return getInt(R.string.pref_key__last_selected_font, 0);
     }
 
     public void setFavoriteMemes(String[] value) {
-        setStringArray(prefApp, R.string.pref_key__meme_favourites, value);
+        setStringArray(R.string.pref_key__meme_favourites, value);
     }
 
     public String[] getFavoriteMemes() {
-        return getStringArray(prefApp, R.string.pref_key__meme_favourites);
+        return getStringArray(R.string.pref_key__meme_favourites);
     }
 
     public void appendFavoriteMeme(String meme) {
@@ -136,15 +136,15 @@ public class AppSettings extends AppSettingsBase {
     }
 
     public void setLastSelectedTab(int value) {
-        setInt(prefApp, R.string.pref_key__last_selected_tab, value);
+        setInt(R.string.pref_key__last_selected_tab, value);
     }
 
     public int getLastSelectedTab() {
-        return getInt(prefApp, R.string.pref_key__last_selected_tab, 0);
+        return getInt(R.string.pref_key__last_selected_tab, 0);
     }
 
     public int getGridColumnCountPortrait() {
-        int count = getInt(prefApp, R.string.pref_key__grid_column_count_portrait, -1);
+        int count = getInt(R.string.pref_key__grid_column_count_portrait, -1);
         if (count == -1) {
             count = 3 + (int) Math.max(0, 0.5 * (Helpers.get().getEstimatedScreenSizeInches() - 5.0));
             setGridColumnCountPortrait(count);
@@ -153,11 +153,11 @@ public class AppSettings extends AppSettingsBase {
     }
 
     public void setGridColumnCountPortrait(int value) {
-        setInt(prefApp, R.string.pref_key__grid_column_count_portrait, value);
+        setInt(R.string.pref_key__grid_column_count_portrait, value);
     }
 
     public int getGridColumnCountLandscape() {
-        int count = getInt(prefApp, R.string.pref_key__grid_column_count_landscape, -1);
+        int count = getInt(R.string.pref_key__grid_column_count_landscape, -1);
         if (count == -1) {
             count = (int) (getGridColumnCountPortrait() * 1.8);
             setGridColumnCountLandscape(count);
@@ -166,20 +166,20 @@ public class AppSettings extends AppSettingsBase {
     }
 
     public void setGridColumnCountLandscape(int value) {
-        setInt(prefApp, R.string.pref_key__grid_column_count_landscape, value);
+        setInt(R.string.pref_key__grid_column_count_landscape, value);
     }
 
     public boolean isAppFirstStart(boolean doSet) {
-        boolean value = getBool(prefApp, R.string.pref_key__app_first_start, true);
+        boolean value = getBool(R.string.pref_key__app_first_start, true);
         if (doSet) {
-            setBool(prefApp, R.string.pref_key__app_first_start, false);
+            setBool(R.string.pref_key__app_first_start, false);
         }
         return value;
     }
 
     public boolean isAppCurrentVersionFirstStart() {
-        int value = getInt(prefApp, R.string.pref_key__app_first_start_current_version, -1);
-        setInt(prefApp, R.string.pref_key__app_first_start_current_version, BuildConfig.VERSION_CODE);
+        int value = getInt(R.string.pref_key__app_first_start_current_version, -1);
+        setInt(R.string.pref_key__app_first_start_current_version, BuildConfig.VERSION_CODE);
         return value != BuildConfig.VERSION_CODE && !BuildConfig.IS_TEST_BUILD;
     }